requirements.txt
ccxt<5
python-dotenv
